feat(linter): auto-fix lint rule violations with `--fix` (#191)
## What This PR Does
Part of #183.

Adds automatic fixing capabilities. Lint rules may now provide a fixer
function when they report an error that, when applied to the linted
source, will fix the violation that was found.

This PR also adds an auto-fixer to `no-catch-return`.

- Testing infrastructure will be added in a follow-up pr. It is blocked
by the upcoming LintService refactor
- Rule docgen still needs to be updated to include fix info in rule
documentation. Again, will be added in an upcoming PR.