fix(linter): multiple file level directives (#195)
👋 hey there, thanks for creating this! i did a quick search and didnt
see this come up in any issues, but apologies if I missed something.

I've got a test file I wanted to do some ignores on like:

```zig
// zlint-disable suppressed-errors, no-undefined
// imports
```

it seemed like only one of the directives was being picked up -- I tried
with and without commas. so I took a look.

I added a simple test following the pattern of existing tests and the
test failed, so did some poking around and made a few tweaks. `just
test` and `just e2e` all seems happy so I think/hope I didnt break
anything here 😄

Changes:
- added test to assert multiple global directives are picked up
- in Parser modified the current cursor position check parse method to
select the min of itself or end of the span (rather than min of itself
+1, end of span)
- this had knock on affect of making some of the span indexes wrong in a
few tests, so I updated the test data there, looking at the test results
I think this branch is still "correct" as it the spans still cover the
full directive (i.e. all of "foo" in the test)
- in src/linter modified the filterDisableRules func to return the view
into the rule_buf from 0 -> `i` of matched rules always - before either
returned this or all configured rules if `i` was 0, which in the case of
the test (or anyone having exactly n:n configured rules to global ignore
directives). this felt a bit scarier, but everything seems to pass/work
still

let me know if this needs any tweaks or anything!

carl

---------

Co-authored-by: Don Isaac <donald.isaac@gmail.com>
Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>